Create Class activity 12
This program calculates the factorial of the positive input. If the input is not a positive number the program would output "Not a positive number" and would ask to insert a positive number. If the input is positive then the number would go through the factorial() function and would calculate the factorial and only return the second to last value before it is equal to zero. After that, the program would then output the input value and then the factorial of the number.