Create class activity 14
This program is ment to generate a random number and would ask the user to guess the number. I had the generator limit the range to 100 so that it can be simpler to guess. After generating the  secret number the program will then run a while loop and ask the user to guess the secret number. If the guess is too low the program would say "too low," and if the guess is too high it would say "Too high." If the user guess figures out the number then the program would output "You got it!!!"