Create Assignment 1
This Program takes a positive input using a void function and outputs the binary representation of the input. In the main function it is used to find the binary as dividing the decimal by 2 and determining the remainder that if the line will have a 1 or 0 in the first line.