Create Assignment 15
This program in C++ creates a column of stars " * " where the program would ask to enter the number of rows and using 'i=1,' if i is less than or equal to the number of rows the user has entered then it goes to the next line where it would ask for a new value called "j" to represent the number of rows. If j is less than or equal to the number of rows then j would increase by 1 until it is greater than the input.