Create Middle Finder
This program allows the user to find the center of a 2 diminutional plain using a series of functions. Using the table function would allow me to organize and clearly display large amounts of information on a grid-type structure with rows and columns so that it would be easier to calculate the center of the x and y axis. the void "findmiddle" function would then calculate the x and y center using floor() which would round the number to the nearest integer. The private function would allow the program to only call the function that may only be called from other functions declared in that class/struct definition.