Create Class Assignment 17
This program locates the targeted value within a array and would output the index in where the value is positioned within said array. Using int linearSearch() allows the program to locate the target value and return a value for the main function to use to output whether it is or isn't the target.