Create Class Assignment
This program uses a void function to sort out the array from smallest to largest value provided using for-loops and pointers. The for loop addresses the position of j within the array size of i. If the array pointer value in position j is greater than position array j+1 then array j will be increased by 1 or (array[j+1]). The program then goes to the main function and with the given array determines the size of such array and calls the void function from before and then prints out both the previous array and the newly sorted one.