Create Class assignment 16
This program helps analyze the grades of a class with having this program it will provide the average, Highest number, lowest number, student that are above average and comparing students grading based on our agreement you have 2 arrays as given. I used the function prototype in order to address the void function so that the program can identify it and make sure to use proceed with it. With the arrays the program uses the double highest as the highest grate and assigns that value as the integer, same thing with the lowest score.