Create Class Assignment 19
This program uses pointer values to swap values between two integers using pointer values. I used a void function so that it can focus on identifying the memory address of the integer and then swap them both using temp as one variable to allow me to swap them. In the main function, I had two values as inputs so that it could use different numbers and then present both values of the inputed integers. Afterward the program then call the swap() function and then outputs the new values of the integers.